+++ package.json
@@ -11,8 +11,11 @@
     "@types/node": "^16.11.46",
     "@types/react": "^18.0.15",
     "@types/react-dom": "^18.0.6",
+    "axios": "^0.27.2",
+    "bootstrap": "^5.1.0",
     "nanoid": "^4.0.0",
     "react": "^18.2.0",
+    "react-bootstrap": "^2.5.0-beta.1",
     "react-confetti": "^6.1.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.3.0",
