--- src/pages/Leaderboard.tsx
@@ -1,9 +1,0 @@
-const Leaderboard = () => {
-    return (
-        <div>
-
-        </div>
-    )
-}
-
-export default Leaderboard
