+++ src/style.css
@@ -3,6 +3,7 @@
 }
 
 header {
+    max-width: 100%;
     height: 50px;
     margin-bottom: 50px;
     color: white;
@@ -10,6 +11,27 @@ header {
     display: flex;
     justify-content: space-between;
     align-items: center;
+}
+
+a {
+    text-decoration: none;
+    color: white;
+}
+
+a:hover {
+    cursor: pointer;
+    text-decoration: underline;
+    text-decoration-color: white;;
+}
+
+.navbar {
+    display: flex;
+    justify-content: space-around;
+    align-items: center;
+}
+
+nav {
+    margin: 20px;
 }
 
 body {
@@ -171,7 +193,7 @@ main {
 
 .win-msg {
     max-width: 80%;
-    margin: 0;
+    margin: auto;
     text-align: center;
     min-height: 50px;
     font-size: 13.1072px;
@@ -186,54 +208,34 @@ main {
     margin: 0;
 }
 
-.form-container {
+form {
     font-family: 'Inter';
-    padding: 20px;
+    padding: 10px;
     display: flex;
-    flex-direction: column;
     justify-content: center;
-    align-items: left;
+    align-items: center;
 }
 
 .form-input {
-    width: 100%;
-    margin-bottom: 20px;
-    height: 40px;
+    margin-inline: 15px;
+    width: 50%;
+    font-size: 15px;
     border-radius: 5px;
     border: 1px solid gray;
-    padding: 5px;
     font-family: 'Inter', sans-serif;
 }
 
 .form-submit {
-    font-size: 100%;
-    width: 70%;
-    padding: 10px;
+    font-family: "Karla";
+    line-height: 19px;
     border-radius: 5px;
     color: white;
-    background-color: #639;
+    background-color: #5035ff;
     border: none;   
     cursor: pointer; 
     align-self: center;
 }
 
-a {
-    text-decoration: none;
-    color: white;
-}
-
-a:hover {
-    cursor: pointer;
-    text-decoration: underline;
-    text-decoration-color: white;;
-}
-
-.navbar {
-    display: flex;
-    justify-content: space-around;
-    align-items: center;
-}
-
-nav {
-    margin: 20px;
+.wrapper {
+    padding-top: 30px;
 }
